fix(exceptions): assembled all exceptions in one file
