+++ Derivas/Expression/Trivial.cs
@@ -2,20 +2,6 @@
 using System;
 using System.Collections.Generic;
 
-namespace Derivas.Exception
-{
-    /// <summary>
-    /// Symbol value not supplied during calculation
-    /// </summary>
-    public class DvSymbolMismatchException : DvBaseException
-    {
-        public string ShouldBe;
-        public DvSymbolMismatchException(string shouldBe)
-            : base($"Value of '{shouldBe}' Symbol is not included in the dictionary")
-            => ShouldBe = shouldBe;
-    }
-}
-
 namespace Derivas.Expression
 {
     internal class Constant : CloneableExpr
