+++ Derivas/Expression/Utils.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Linq;
 using System.Collections.Generic;
-using System.Runtime.CompilerServices;
 
 using Derivas.Exception;
